requirements.txt
python-dotenv>=1.0.0
Flask>=2.3.0
Flask-WTF>=1.2.0
Flask-Limiter>=3.5.0
Flask-Login>=0.6.3
Flask-Mail>=0.9.1
Flask-SocketIO>=5.3.0
python-socketio>=5.10.0
bleach>=6.1.0
reportlab>=4.0.0
qrcode>=7.4.2
python-barcode>=0.15.1
Werkzeug>=2.3.0
xhtml2pdf>=0.2.15
openpyxl>=3.1.0
pytest>=7.4.0
pytest-playwright>=0.4.0
playwright>=1.40.0

# Production
gunicorn>=21.0.0
psycopg2-binary>=2.9.9
